Fix more warnings (#1005)
we have too much of them and our compilation log is too huge

+++ src/fheroes2/gui/text.h
@@ -146,11 +146,11 @@ public:
 
     int w( void ) const
     {
-        return gw;
+        return static_cast<int>( gw );
     }
     int h( void ) const
     {
-        return gh;
+        return static_cast<int>( gh );
     }
 
     void Blit( s32, s32, Surface & sf = Display::Get() ) const;
